Restructure blueprints to be in blueprints directory and integrate orphaned blueprints from collective.blueprint.jsonmigrator.
